Add some documentation/specification
Some of it seems to be broken right now, but
the first step was saying what it was supposed to do.
Then I can write tests/fixes better.

Anyway it needed a good help string; I always forgot
what my own program's syntax was on the occasions I wanted
to use it.