Use a more perfect random algorithm.
Now, between this and urandom,
the random numbers are Not Obviously Risky
(though I didn't double-check thoroughly enough
to be sure it's all cryptographically sound!
Even aside from timing attacks which are
very likely possible against this implementation
if that's part of your threat model. etc :-) )

Note that the bit-selecting method I chose
is actually a bad idea for some RNGs whose
low-order bits are less good. Perhaps I'll
change it, or perhaps add a method
rand_ops.rand_n_bits(n) for rands to optionally
implement.  Note that the libc_rand implementation
actually calls rand() a few times and shifts
around the bits in a way that this is less
of a problem (libc_rand_val's low order bits
are an xor of rand()'s low order and middle
bits, more or less).  So it's not as bad as
HORRORS!!!, which is probably as good as you
were looking for if you're going for libc rand.

Of course I know I'm probably reimplementing
a library here that already exists! But it's
fun and I got something I wanted pretty efficiently
and I got to learn about creating C librariness
( also cf. https://lwn.net/Articles/444910/ )
- I might indeed swap in a real existing library
if I ever feel the need/desire.  Or you could!!