Remember to add a license (permissive MIT)
This is a small enough amount of code that it seems rather
pointless, and annoying to users of the code, to apply the GPL.

I very well might change to GPL if somehow the code burgeons
by a factor of ten or so.  Which I doubt will happen.  In any
case, this historical version would always be MIT-licensed,
because that's how free/libre/open-source copyrights work.